+++ petitRADTRANS/retrieval/data.py
@@ -5,9 +5,9 @@ import warnings
 
 import numpy as np
 from astropy.io import fits
-from petitRADTRANS.fortran_convolve import fortran_convolve as fconvolve
 from petitRADTRANS.fortran_rebin import fortran_rebin as frebin
-from scipy.ndimage import gaussian_filter
+from petitRADTRANS.math import convolve, convolve_and_sample_Rvers, filter_spectrum_with_spline
+from spectres.spectral_resampling_numba import spectres_numba
 
 import petitRADTRANS.physical_constants as cst
 
@@ -126,6 +126,8 @@ class Data:
                  scale=False,
                  scale_err=False,
                  offset_bool=False,
+                 resample=False,
+                 subtract_continuum=False,
                  wavelength_bin_widths=None,
                  photometry=False,
                  photometric_transformation_function=None,
@@ -197,6 +199,8 @@ class Data:
         self.scale = scale
         self.scale_err = scale_err
         self.offset_bool = offset_bool
+        self.resample = resample
+        self.subtract_continuum = subtract_continuum
         self.scale_factor = 1.0
         self.offset = 0.0
         self.bval = -np.inf
@@ -485,13 +489,13 @@ class Data:
 
                 for spectrum_model in model_spectra:
                     if self.data_resolution_array_model is not None:
-                        spectrum_model = self.convolve(
+                        spectrum_model = convolve(
                             wlen_model,
                             spectrum_model,
                             self.data_resolution_array_model
                         )
                     elif self.data_resolution is not None:
-                        spectrum_model = self.convolve(
+                        spectrum_model = convolve(
                             wlen_model,
                             spectrum_model,
                             self.data_resolution
@@ -507,6 +511,25 @@ class Data:
                     else:
                         rebin = True
 
+                    if self.name + "_radial_velocity" in parameters.keys():
+                        # RV in km/s -> multiply by 1e5 to cm/s
+                        # wlen_model in micron
+                        # cst.c in cm/s
+                        radial_velocity = parameters[self.name + "_radial_velocity"].value * 1e5 
+                        wlen_model *= wlen_model * np.sqrt((1 + radial_velocity/cst.c)/(1- radial_velocity/cst.c)) 
+                        #wlen_model += wavel_shift
+
+                        # Don't want to rebin the spectrum multiple times.
+                        #spectrum_model = spectres_numba(
+                        #    wlen_model,
+                        #    wlen_model + wavel_shift,
+                        #    verbose=False
+                        #    spectrum_model,
+                        #)
+                    elif "system_radial_velocity" in parameters.keys():
+                        radial_velocity = parameters["system_radial_velocity"].value * 1e5 
+                        wlen_model *= np.sqrt((1 + radial_velocity/cst.c)/(1 - radial_velocity/cst.c)) 
+                        #wlen_model += wavel_shift
                     if rebin:
                         flux_rebinned = frebin.rebin_spectrum_bin(
                             wlen_model,
@@ -515,6 +538,11 @@ class Data:
                             self.wavelength_bin_widths
                         )
 
+                    if self.atmospheric_column_flux_mixer is None:
+                        break
+                    else:
+                        column_rebinned_spectra.append(flux_rebinned)
+                    
                     if self.atmospheric_column_flux_mixer is None:
                         break
                     else:
@@ -532,6 +560,19 @@ class Data:
             # species spectrum_to_flux functions return (flux,error)
             if isinstance(flux_rebinned, (tuple, list)):
                 flux_rebinned = flux_rebinned[0]
+
+        if self.resample:
+            r_slope = parameters[self.name + "_R_slope"].value
+            r_intersect = parameters[self.name + "_R_int"].value
+            r_array = (self.wavelengths*r_slope)+r_intersect
+            # TODO: interpolate r_array and use standard convolve function.
+            # TODO: compare different convolution and binning methods
+            flux_rebinned = convolve_and_sample_Rvers(self.wavelengths, r_array, self.wavelengths, flux_rebinned)
+        
+        if self.subtract_continuum:
+            nodes = parameters[self.name + "_nodes"].value
+            x_nodes = np.linspace(self.wavelengths[0], self.wavelengths[-1], nodes)
+            flux_rebinned = filter_spectrum_with_spline(self.wavelengths,flux_rebinned,x_nodes=x_nodes)
 
         if self.scale:
             diff = (flux_rebinned - self.spectrum * parameters[self.name + "_scale_factor"].value) + self.offset
@@ -751,44 +792,3 @@ class Data:
             elif 'uncertainty_scaling_b' in parameters.keys():
                 b_val = parameters['uncertainty_scaling_b'].value
         return b_val
-
-    @staticmethod
-    def convolve(input_wavelength,
-                 input_flux,
-                 instrument_res):
-        r"""
-        This function convolves a model spectrum to the instrumental wavelength
-        using the provided data_resolution
-        Args:
-            input_wavelength : numpy.ndarray
-                The wavelength grid of the model spectrum
-            input_flux : numpy.ndarray
-                The flux as computed by the model
-            instrument_res : float
-                :math:`\\lambda/\\Delta \\lambda`, the width of the gaussian kernel to convolve with the model spectrum.
-
-        Returns:
-            flux_lsf
-                The convolved spectrum.
-        """
-        if isinstance(instrument_res, np.ndarray):
-            return fconvolve.variable_width_convolution(input_wavelength, input_flux, instrument_res)
-        # From talking to Ignas: delta lambda of resolution element
-        # is FWHM of the LSF's standard deviation, hence:
-        sigma_lsf = 1. / instrument_res / (2. * np.sqrt(2. * np.log(2.)))
-
-        # The input spacing of petitRADTRANS is 1e3, but just compute
-        # it to be sure, or more versatile in the future.
-        # Also, we have a log-spaced grid, so the spacing is constant
-        # as a function of wavelength
-        spacing = np.mean(2. * np.diff(input_wavelength) / (input_wavelength[1:] + input_wavelength[:-1]))
-
-        # Calculate the sigma to be used in the gauss filter in units
-        # of input wavelength bins
-        sigma_lsf_gauss_filter = sigma_lsf / spacing
-
-        flux_lsf = gaussian_filter(input_flux,
-                                   sigma=sigma_lsf_gauss_filter,
-                                   mode='nearest')
-
-        return flux_lsf