+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to petitRADTRANS will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com)
 and this project adheres to [Semantic Versioning](http://semver.org).
 
+## [3.3.0a0] - 2025-05-22
+### Added
+- Linear spline continuum subtraction (courtesy Will Balmer)
+- Radial velocity handling in retrieval package (courtesy Will Balmer)
+
 ## [3.2.0a31] - 2025-05-15
 ### Added
 - Non-vertically constant free-chemistry abundance profiles. Included stepped, linear and cubic interpolations.